+++ changelog.md
@@ -5,8 +5,8 @@
 ### Added
 * Support `page-range:*` to select changed pages when using `watch`
 * Add `spread-percent`
-* Add shortcuts for the values of `distribute-rows`, `distribute-columns`, and `distribute-items` 
- 
+* Add shortcuts for the values of `distribute-rows`, `distribute-columns`, and `distribute-items`  on page, row, column item types
+* Add shortcuts for row-gutter, column-gutter, item-gutter on page, row, column item types
 ### Changed
 * Default `watch` to true, but do not watch if there is nothing to watch
 * Handle multiple books by merging into one